Move TrailingCommaRule from Experimental to Standard ruleset
