+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -2,7 +2,6 @@ package com.pinterest.ktlint.ruleset.experimental
 
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
-import com.pinterest.ktlint.ruleset.experimental.trailingcomma.TrailingCommaRule
 
 public const val experimentalRulesetId = "experimental"
 
@@ -13,7 +12,6 @@ public class ExperimentalRuleSetProvider : RuleSetProvider {
         ArgumentListWrappingRule(),
         MultiLineIfElseRule(),
         NoEmptyFirstLineInMethodBlockRule(),
-        TrailingCommaRule(),
         PackageNameRule(),
         EnumEntryNameCaseRule(),
         SpacingAroundDoubleColonRule(),
