+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/StandardRuleSetProvider.kt
@@ -2,6 +2,7 @@ package com.pinterest.ktlint.ruleset.standard
 
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
+import com.pinterest.ktlint.ruleset.experimental.trailingcomma.TrailingCommaRule
 
 public class StandardRuleSetProvider : RuleSetProvider {
 
@@ -38,6 +39,7 @@ public class StandardRuleSetProvider : RuleSetProvider {
         SpacingAroundParensRule(),
         SpacingAroundRangeOperatorRule(),
         StringTemplateRule(),
+        TrailingCommaRule(),
         WrappingRule()
     )
 }
